Introduce GitHub App provider
This also splits out the common logic between the GitHub app implementation
and the GitHub OAuth implementation.

Ref #2520
