+++ internal/projects/create.go
@@ -27,7 +27,7 @@ import (
 
 	"github.com/stacklok/minder/internal/authz"
 	"github.com/stacklok/minder/internal/db"
-	github "github.com/stacklok/minder/internal/providers/github"
+	github "github.com/stacklok/minder/internal/providers/github/oauth"
 	pb "github.com/stacklok/minder/pkg/api/protobuf/go/minder/v1"
 )
