+++ internal/providers/github/oauth/oauth_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package github
+package oauth
 
 import (
 	"context"
@@ -29,6 +29,7 @@ import (
 
 	"github.com/stacklok/minder/internal/db"
 	"github.com/stacklok/minder/internal/providers/credentials"
+	github2 "github.com/stacklok/minder/internal/providers/github"
 	"github.com/stacklok/minder/internal/providers/ratecache"
 	provtelemetry "github.com/stacklok/minder/internal/providers/telemetry"
 	minderv1 "github.com/stacklok/minder/pkg/api/protobuf/go/minder/v1"
@@ -53,7 +54,7 @@ func TestArtifactAPIEscapes(t *testing.T) {
 	tests := []struct {
 		name        string
 		testHandler http.HandlerFunc
-		cliFn       func(cli *GitHub)
+		cliFn       func(cli *github2.GitHub)
 		wantErr     bool
 	}{
 		{
@@ -62,7 +63,7 @@ func TestArtifactAPIEscapes(t *testing.T) {
 				assert.Equal(t, "/orgs/stacklok/packages/container/helm%2Fmediator", r.URL.RequestURI())
 				w.WriteHeader(http.StatusOK)
 			},
-			cliFn: func(cli *GitHub) {
+			cliFn: func(cli *github2.GitHub) {
 				_, err := cli.GetPackageByName(context.Background(), true, "stacklok", "container", "helm/mediator")
 				assert.NoError(t, err)
 			},
@@ -73,7 +74,7 @@ func TestArtifactAPIEscapes(t *testing.T) {
 				assert.Equal(t, "/orgs/stacklok/packages/container/helm%2Fmediator/versions?package_type=container&page=1&per_page=100&state=active", r.URL.RequestURI())
 				w.WriteHeader(http.StatusOK)
 			},
-			cliFn: func(cli *GitHub) {
+			cliFn: func(cli *github2.GitHub) {
 				_, err := cli.GetPackageVersions(context.Background(), true, "stacklok", "container", "helm/mediator")
 				assert.NoError(t, err)
 			},
@@ -84,7 +85,7 @@ func TestArtifactAPIEscapes(t *testing.T) {
 				assert.Equal(t, "/orgs/stacklok/packages/container/helm%2Fmediator/versions?package_type=container&page=1&per_page=100&state=active", r.URL.RequestURI())
 				w.WriteHeader(http.StatusOK)
 			},
-			cliFn: func(cli *GitHub) {
+			cliFn: func(cli *github2.GitHub) {
 				_, err := cli.GetPackageVersionByTag(context.Background(), true, "stacklok", "container", "helm/mediator", "v1.0.0")
 				assert.NoError(t, err)
 			},
@@ -95,7 +96,7 @@ func TestArtifactAPIEscapes(t *testing.T) {
 				assert.Equal(t, "/orgs/stacklok/packages/container/helm%2Fmediator/versions/123", r.URL.RequestURI())
 				w.WriteHeader(http.StatusOK)
 			},
-			cliFn: func(cli *GitHub) {
+			cliFn: func(cli *github2.GitHub) {
 				_, err := cli.GetPackageVersionById(context.Background(), true, "stacklok", "container", "helm/mediator", 123)
 				assert.NoError(t, err)
 			},
@@ -208,7 +209,7 @@ func TestConcurrentWaitForRateLimitReset(t *testing.T) {
 			client, ok := restClientCache.Get(owner, token, db.ProviderTypeGithub)
 			require.True(t, ok)
 
-			ghClient, ok := client.(*GitHub)
+			ghClient, ok := client.(*github2.GitHub)
 			require.True(t, ok)
 
 			_, err := ghClient.CreateIssueComment(ctx, owner, "mockRepo", 1, "Test Comment")
